Fix blocked processes not resuming, download other trace files, small changes to stat

+++ stat.h
@@ -10,10 +10,10 @@
 // The tmu and trp fields get converted into the correct amu and arp fields at the
 // point of the program exit, where they are divided by the total time.
 struct stat_t {
-    long tmu; // total memory utilization (compounding)
-    long trp; // total runable processes (compounding)
-    long tmr; // total memory references
-    long tpi; // total page ins: number of page faults
+    float tmu; // total memory utilization (compounding)
+    float trp; // total runable processes (compounding)
+    float tmr; // total memory references
+    float tpi; // total page ins: number of page faults
 };
 
 // Initialize stat structure
